refactor: organize services by function instead of type
