+++ src/Services/Download/AbstractDownloadService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace AspirePress\AspireSync\Services;
+namespace AspirePress\AspireSync\Services\Download;
 
 use AspirePress\AspireSync\Services\Interfaces\MetadataServiceInterface;
 use Exception;
