+++ src/Services/List/PluginListService.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace AspirePress\AspireSync\Services\Plugins;
+namespace AspirePress\AspireSync\Services\List;
 
-use AspirePress\AspireSync\Services\AbstractListService;
+use AspirePress\AspireSync\Services\List\AbstractListService;
 use AspirePress\AspireSync\Services\Interfaces\SubversionServiceInterface;
+use AspirePress\AspireSync\Services\Metadata\PluginMetadataService;
 use AspirePress\AspireSync\Services\RevisionMetadataService;
 
 readonly class PluginListService extends AbstractListService
