+++ src/Services/List/ThemeListService.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace AspirePress\AspireSync\Services\Themes;
+namespace AspirePress\AspireSync\Services\List;
 
-use AspirePress\AspireSync\Services\AbstractListService;
+use AspirePress\AspireSync\Services\List\AbstractListService;
 use AspirePress\AspireSync\Services\RevisionMetadataService;
 use AspirePress\AspireSync\Services\SubversionService;
+use AspirePress\AspireSync\Services\Metadata\ThemeMetadataService;
 
 readonly class ThemeListService extends AbstractListService
 {
